reorgonized 1/2 auto 1 path fix linear slide
