+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/PowerPlayTeleOp.java
@@ -27,6 +27,9 @@ public class PowerPlayTeleOp extends LinearOpMode {
     final double scissorClosed = 0.7;
     final double scissorOpen = 0;
     double scissorPosition = scissorClosed;
+    final int manualSlideOn = 1;
+    final int manualSlideOff = 0;
+    int slideToggle = manualSlideOff;
     Servo left_servo;
 
     static final int CYCLE_MS =  500; // period of each cycle
@@ -128,9 +131,22 @@ public class PowerPlayTeleOp extends LinearOpMode {
     }
 
     private void processLinearSlide(){
+        if (gamepad1.left_bumper &&(timeSinceLastPress.milliseconds() >= BUTTON_DELAY)){
+            if (slideToggle == manualSlideOff){
+                slideToggle = manualSlideOn;
+                telemetry.addData("Manual Slide", "ON");
+            }
+            else{
+                slideToggle = manualSlideOff;
+                telemetry.addData("Manual Slide", "OFF");
+            }
+
+            timeSinceLastPress.reset();
+        }
         double y = gamepad2.left_stick_y;
-        leftLinearSlide.setPower(y / 2);
-        rightLinearSlide.setPower(y / 2);
+        leftLinearSlide.setPower(y/2);
+        rightLinearSlide.setPower(y/2);
+
 
         telemetry.addData("Left Linear Slide Position", leftLinearSlide.getCurrentPosition());
         telemetry.addData("Right Linear Slide Position", rightLinearSlide.getCurrentPosition());
@@ -141,15 +157,22 @@ public class PowerPlayTeleOp extends LinearOpMode {
         if (gamepad1.y && (timeSinceLastPress.milliseconds() >= BUTTON_DELAY)) {
             leftLinearSlide.setTargetPosition(1840);
             rightLinearSlide.setTargetPosition(1803);
+            moveSlides();
+
         }
         if (gamepad1.a && (timeSinceLastPress.milliseconds() >= BUTTON_DELAY)) {
             leftLinearSlide.setTargetPosition(3048);
             rightLinearSlide.setTargetPosition(3030);
+            moveSlides();
         }
         if (gamepad1.b && (timeSinceLastPress.milliseconds() >= BUTTON_DELAY)) {
             leftLinearSlide.setTargetPosition(2013);
             rightLinearSlide.setTargetPosition(2008);
+            moveSlides();
         }
+    }
+
+    private void moveSlides(){
         leftLinearSlide.setMode(DcMotor.RunMode.RUN_TO_POSITION);
         rightLinearSlide.setMode(DcMotor.RunMode.RUN_TO_POSITION);
         leftLinearSlide.setPower(0.15);
@@ -174,11 +197,9 @@ public class PowerPlayTeleOp extends LinearOpMode {
         }
 
         // Display the current value
-        telemetry.addData("X Position", "%5.2f", scissorClosed);
-        telemetry.addData("Y Position", "%5.2f", scissorOpen);
-        telemetry.addData("Y Position is open", "");
-        telemetry.addData("X Position is closed", "");
-        telemetry.addData("Actual Servo Position is", "%5.2f", left_servo.getPosition());
+        telemetry.addData("Scissor constants",  "(open=%5.2f, closed=%5.2f)", scissorOpen, scissorClosed);
+        telemetry.addData("Desired Scissor Position", scissorPosition);
+        telemetry.addData("Actual Scissor Position", "%5.2f", left_servo.getPosition());
     }
 }
