+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/old/Driving.java
@@ -1,6 +1,7 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.old;
 
 import com.qualcomm.hardware.bosch.BNO055IMU;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.DcMotor;
@@ -15,6 +16,7 @@ import org.firstinspires.ftc.robotcore.external.navigation.AxesReference;
 import java.math.RoundingMode;
 import java.text.DecimalFormat;
 
+@Disabled
 @TeleOp
 public class Driving extends LinearOpMode {
     static final double INCREMENT   = 0.05;
